% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsdmx.R
\name{read_sdmx}
\alias{read_sdmx}
\title{Read SDMX data}
\usage{
read_sdmx(path)
}
\arguments{
\item{path}{the path to an SDMX-ML xml file or an url to connect to.}
}
\value{
data.frame object
}
\description{
read SDMX data into R as dataframes
}
\details{
read_sdmx returns a dataframe with the SDMX data 'as-is'
which means that all columns will be character vectors.
}
